fixing lint and build errors

+++ src/nvim/ex_cmds2.c
@@ -2806,8 +2806,8 @@ static char_u *get_str_line(int c, void *cookie, int ident)
   char buf[2046];
   char *dst;
   dst = xstpncpy(buf, (char *)p->buf+p->pointer, i);
-  if(dst - buf != i){
-	smsg(_("nvim_source error parsing command %s"),p->buf);
+  if ((uint32_t)(dst - buf) != i) {
+    smsg(_("nvim_source error parsing command %s"), p->buf);
   }
   buf[i-p->pointer]='\0';
   p->pointer = i + 1;
